Use the ClassLoader class of composer extension repository
The used ClassLoader in library is outdated (API changes). It will cause an error when using the NamespaceClassLoader in an Contao setup where the Composer extension repository is used because the ClassLoader of this extension is used for the whole system.

Updating the ClassLoader would solve this issue. I think it is better to use the ClassLoader of the main composer installation if available